requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.6.0
anyio==3.7.1
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
audioread==3.0.1
backcall==0.2.0
blinker==1.6.2
cachetools==5.3.3
certifi==2023.5.7
cffi==1.16.0
chardet==3.0.4
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
comm==0.1.3
comtypes==1.3.1
contextlib2==21.6.0
cryptography==44.0.2
dataclasses-json==0.6.4
datasets==2.19.1
debugpy==1.6.7
decorator==5.1.1
dill==0.3.8
dnspython==2.3.0
exceptiongroup==1.2.1
execnet==2.1.1
executing==1.2.0
faiss-cpu==1.8.0
filelock==3.14.0
Flask==2.3.2
Flask-PyMongo==2.3.0
frozenlist==1.3.3
fsspec==2024.3.1
gitdb==4.0.11
GitPython==3.1.43
googletrans==4.0.0rc1
greenlet==3.0.3
gTTS==2.5.1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.4.1
httpcore==0.9.0
httpx==0.13.3
huggingface-hub==0.30.1
hyperframe==5.2.0
idna==2.10
iniconfig==2.0.0
ipykernel==6.23.2
ipython==8.14.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jiwer==2.6.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.2.0
jupyter_core==5.3.0
langchain==0.1.14
langchain-community==0.0.30
langchain-core==0.1.37
langchain-text-splitters==0.0.1
langsmith==0.1.38
librosa==0.9.2
llvmlite==0.42.0
lxml==5.3.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.21.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mock==5.1.0
more-itertools==10.6.0
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.16
mypy-extensions==1.0.0
nest-asyncio==1.5.6
networkx==3.3
numba==0.59.1
numpy==1.26.4
openai==0.27.8
openai-whisper==20240930
orjson==3.10.0
packaging==23.2
pandas==2.2.1
parso==0.8.3
path==16.14.0
path.py==12.5.0
pdfminer.six==20250327
pdfplumber==0.11.6
pickleshare==0.7.5
pillow==10.2.0
platformdirs==3.5.3
pluggy==1.5.0
pooch==1.8.1
prompt-toolkit==3.0.38
protobuf==4.25.3
psutil==5.9.5
pure-eval==0.2.2
pyarrow==15.0.2
pyarrow-hotfix==0.6
pycparser==2.22
pydantic==2.6.4
pydantic_core==2.16.3
pydeck==0.8.1b0
Pygments==2.17.2
pymongo==4.4.0
PyPDF2==3.0.1
pypdfium2==4.30.1
# pypiwin32==223
pytesseract==0.3.10
pytest==8.2.0
pytest-shutil==1.7.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
pyttsx3==2.90
pytube==15.0.0
pytz==2024.1
# pywin32==306
PyYAML==6.0.1
pyzmq==25.1.0
rapidfuzz==2.13.7
referencing==0.34.0
regex==2023.12.25
requests==2.31.0
resampy==0.4.3
rfc3986==1.5.0
rich==13.7.1
rpds-py==0.18.0
safetensors==0.4.3
scikit-learn==1.4.2
scipy==1.13.0
sentence-transformers==4.0.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soundfile==0.12.1
SpeechRecognition==3.10.3
SQLAlchemy==2.0.29
stack-data==0.6.2
streamlit==1.32.2
streamlit-chat==0.1.1
sympy==1.13.1
tenacity==8.2.3
termcolor==2.4.0
threadpoolctl==3.5.0
tiktoken==0.6.0
tokenizers==0.21.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.1
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tornado==6.4
tqdm==4.65.0
traitlets==5.9.0
transformers==4.50.3
typing-inspect==0.9.0
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.0.3
watchdog==4.0.0
wcwidth==0.2.6
Werkzeug==2.3.6
xxhash==3.4.1
yarl==1.9.2
yt-dlp==2025.3.31
